combining Stake.sol & CreditProtocol.sol
- updating package.json to no longer depend on aupiff's fork of
`tce-contracts`
- updaing package-lock.json
