+++ package.json
@@ -24,7 +24,7 @@
     "chai-as-promised": "^7.1.1",
     "chai-bignumber": "^2.0.1",
     "ethereumjs-testrpc": "4.1.1",
-    "tce-contracts": "git@github.com:aupiff/tce-contracts.git#92de2e06e55581e3d3a6c9fcd2a2bd30a53c4ff5",
+    "tce-contracts": "git@github.com:blockmason/tce-contracts.git#92de2e06e55581e3d3a6c9fcd2a2bd30a53c4ff5",
     "truffle": "3.4.9",
     "zeppelin-solidity": "1.3.0"
   }
